archive/streamlit_v1_dashboard/dashboard/requirements_dashboard.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=1.5.0
python-dotenv>=1.0.0
